Minor UI Updates
Made labels  shorter on the 3D View tab so that labels aren't cut off in default panel width.

Also added many labels/descriptions to properties and buttons so that when you look at the add-on at first glance, you see that all the minor details are attended to. Basic goal is for everything to look no less professional as the built-in Blender UI. That means minimal missing operator descriptions and no missing labels that allow "something.some_operator" or "some_property_name_in_snake_case" to be shown in the UI on tooltips.