Fix for CPVIA System
This fixes the problem that arose out of the int_manual_fixture_id or whatever thing being deleted. That Idiot Part is now completely gone and CPVIA system at least seems to be working. I know I broke something though, but haven't found it yet. May have been on the button at the bottom of the Patch Groups panel?