UI Fix
Apparently there is some heisenbug inside Blender that is preventing a specific instance of Blender on my main computer from clearing its cache of the icon files, which were originally mixed up. I fixed the files themselves, but Blender refused to update its internal cache, so I briefly comitted the fix for my machine just to speed up testing on the other machine, which proved the problem was just with my computer. This commit fixes it for everyone but me and breaks it for me.