Duplicate Strips Ups Fixture #
Allowed the duplicate strip in Sequencer toolbar to also increase the channel of duplicated strip, meaning the fixture channel. This is for offset animations.

Also removed the problematic LIGHTPROBE_VOLUME icon from nodes for compatibility with lower versions, since someone at Blender decided to change the icon name between 4.0 and 4.1