Patch Groups Fix
The euler stuff got messed up. This fixes it. I was trying to get it to consider modifier changes, but couldn't get it to work right. I reverted it back to a middle version in the last commit, not realizing that not even the middle version worked right. Reverting back to the original method that actually works for 90% of use cases.