Refactored Influencer Logic for Speed Improvement
FPS with an influencer in the scene along with hundreds of other objects was bringing FPS down to like 1 fps. I realized this was at least partially because the is_inside_mesh function was needlessly calculating position data for the parent object every single iteration for the same parent. I refactored that to put that logic outside of the loop so it only runs once. This resulted in an increase from about 1 fps to 9-10 fps.

Will investigate further later for additional oppurtunities for improvement.

I also added numerous time print statements to the Service Mode debug system. This made me realize that we are running a brand new CPVIA for the same parent multiple times if the parent has multiple properties. We should probably find a way some time to make the CPVIA only run once per parent. That will require some pretty hefty changes to how CPVIA works though.