Centering Fix for Mixer
Trying to keep these fixes separated so we can go back if one breaks stuff and so we can keep track of exactly what changes are being made in which commits.

This commit fixes an issue where the mixer is not properly centered.

Another fix should be coming soon that should fix unexpected behavior that happens when the length of the channels is less than the length of value choices. In those cases it doesn't quite handle the odd situation the way we want. But that fix is coming later.