Fix for Mixer
Alright, this is the best version I can come up with at the time being. Fixing the centering issue by changing endpoint from False to True ending up adding a super undesirable wrapping issue, so we're going back to False. But this commit does fix the issue where everything was wonky when the number of channels was greater than the number of values. That's working pretty well now, although it as well is not centered.

Also looking at an issue where the subdivisions create an unexpected decrease in values overall. I guess it might make sense that this would happen from a maths standpoint, but that's not relevant to the end user. It should work as expected regardless of the underlying maths.