Switched Add Node Ops to Built-in Op
Allowing transform ability for node operators by using Blender's built-in node.add_node operator.

I found that operator in the source and it's actually a Python operator. All the logic for the dynamic transform effect is right there in Python. So I wonder if we can copy that logic, make some modifications, and make it work for Sequencer? Or maybe that's already built-in and you just have to enable it somewhere, like how in 3D view you can select different options for where new stuff should go.