Bug Fix: Very temporary fix for wm right click menu UI draw
This was printing a runtime error to terminal, so I just temporarily disabled it for the Beta release since it's a known bug that I don't want to actually fix right this moment. Will make this stuff work for nodes later but right now it doesn't and that's fine for the Beta.