CPVIA: Set user 0, only for CPVIA and events
Adding a hardcoded user system (the ETC Eos meaning of "user") so that CPVIA and Event Manager operations are on user 0 and everything else defaults to user 1.

Later, this will be implemented into the UI so that the software-user can change this.

This prevents the command line and yellow highlight in A3D from going crazy during playback, but it still allows you to see what you're doing for other operations.