CPVIA: Allow mixer animation
This allows user to use group controller nodes to animate/drive mixer node parameters.

This is a workaround for the issue where Blender's animation system can't keyframe PropertyGroup properties directly.

It adds a Lighting Input socket to the mixer. The order is determined by which node was attached to the socket first. The input link limit is arbitrarily set to 10.