CPVIA: Allow strobe for simpler fixtures like Intimidators
This allows fixtures that enable/disable strobe with a Shutter_Strobe command to be controlled by Sorcerer. Previously, the value command and enable/disable command would conflict. Now, if value is zero, the value command is stripped and we just send the disable command.

This seems to work okay for both fixture types, those that use a dedicated Strobe_Mode parameter as well as those with a single parameter.