Upgrade: Expand limited support to grandMA3
Sorcerer now has limited support for grandMA3!!!

I've disabled 99% of features that are not yet supported via the UI. One feature I know that still remains in the UI that doesn't work yet for MA is the highlight feature for the group maker.

Will later add an updater or something or maybe secondary text properties so that the defaults for gobo stuff are suitable for MA.

I have no idea how to get strobe to work yet, but you may be able to get it to work as end user. Haven't truly dove into it yet.