Delete: Deleting Group System & Param Toggles for Ob
The best part is no part. Can't have bugs, can't be confusing, and can't contribute to learning curve. The textual input is so good that there was exactly 1 node in Hamster Dance that didn't use it. Nothing else in the entire project used it. So we're deleting it.

If we have to add it back in, that's ok.

Also deleting the param toggle from object controller since they don't need to be there. There's really no case where you need it because of how these are set up. Need to also delete from Sequencer. But keep for nodes obv since that UI system benefits from this type of optimization.

Also, this helps us eventually delete the entire side-panel-based Object controller. We want to do that because the side panel wastes a lot of space when it is not being used. You can collapse it but it's not optimized for 3D viewport the way the header bar is.