3D Audio: Irregular Sound Object Scaling Improvement
Now handles sound objects with asymnetric scaling better. Currently, user must add vertices if a speaker is in the middle of a mesh for it to hit correctly. This type of thing is great for beds.

In Atmos, sound objects and beds are totally different things. In PanLab I don't think you have beds at all.

With this, you can think of a "bed" as just a large 3D audio object. This is a far simpler, more first-principlesy way to think about it. "Bed" is just a made up word that is there only because Atmos can't do asymnetric scaling. Sorcerer can.