3D Gradients: Partial Implementation
This is still a bit buggy. Use the latest official release for real projects.

This commit allows meshes within a user-created collection to essentially create 3D gradients. It's sort of like normal influencers, but with feather.

Sorcerer will in the future have its own section in Preferences to enable/disable experimental, early  features like this one.

Old .blend files before this commit are currently throwing errors from this new code, this will be addressed soon. Versions of Sorcerer after this feature addition will be able to work with influencers created before, but behavior may not be as expected. Will not be forwards compatible though, meaning older Sorcerers will not be able to read the newer files correctly. That's because the newer UI draw omits key controls that are now deleted.