Fix: Repair Noodle Finding Logic
This was broken for a long time. It isn't perfect yet, but it's much better than before. Mixers are currently already down in the latest versions of Blender that I'm currently building, so haven't properly tested on mixers yet. I will later expand this logic to handle reroutes as well, since those don't work yet on this commit.