spy: Add .explain() to spy Function
In command line or Text Editor, type spy.make_eos_macros.explain() to get it to explain how to call the function and what it does.

This also fixes an issue where it didn't work at all because of the last major big deal commit that accidentally broke with an improper indent/refactor.

I also experimented with transferring this over to TCP, but it did not immediately work.