WIP: Rewriting all Orb Logic
This is the much-needed rewrite of Orb. This will utilize the new TCP protocol to significantly speed up Orb speed. It also greatly enhances the readability, maintainability, and extendability of the system. It does that by shrinking the footprint of non-shared logic. The non-shared portion is now as small as it can possibly be. This makes the end logic itself much more readable and easy to jump into.