Refactor: Move Discreet Time Logic to Orb.py not Console
Goal is to move as much logic as possible away from the console onto the orb.py. That way, only the logic that is truly console-specific is on the console data class. Thereby, if you want to add Orb support to a new console like MA, the workload is very small.