Macro Strips: Fix Orb and Update Old Logic
Clarifies some very old function naming (this was original Sequencer code I think!) and improved the logic for better handling.

If user inputs something like "Go to cue out", it will NOT add enter to the end.

But, if user inputs something like "Go to cue 1", it WILL add enter to the end.

I also fixed an issue where it would add Enter to the end if the user typed lowercase "enter" at the end.