Fix: Pan/Tilt Node Was Broke Again
On the light I'm working with now, the mapping is wrong too. Will investigate further later.

Every time the CPV system gets rewritten, I forget to test this stupid node (or decide not too because I'm lazy), and as a result, it's basically always broken because CPV gets rewritten every moon cycle.