CPV: Structurally Allow Batching Fix
This does not add batching, but restructures the event manager and publisher so as to allow that in a later commit. Previously, OSC was sent inside the publisher, which runs on a per-request basis. So there was no way to batch OSC commands. Now, the OSC transmit step has been put in the event manager, and only when it is already harmonized, which is only true in the case it is called by event_manager.